+++ test/unit/fuzz-tools.test.js
@@ -138,6 +138,8 @@ import * as idV137 from '../../lib/id-v137.js';
 import * as obV138 from '../../lib/ob-v138.js';
 import * as gynV139 from '../../lib/gyn-v139.js';
 import * as pedsV140 from '../../lib/peds-v140.js';
+import * as pedsGrowthV141 from '../../lib/peds-growth-v141.js';
+import * as growthLmsData from '../../lib/growth-lms-data.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -221,6 +223,8 @@ const MODULES = {
   'ob-v138.js': obV138,
   'gyn-v139.js': gynV139,
   'peds-v140.js': pedsV140,
+  'peds-growth-v141.js': pedsGrowthV141,
+  'growth-lms-data.js': growthLmsData,
   'ems-v149.js': emsV149,
 };